requirements.txt
streamlit
plotly
numpy
pandas
openai
python-dotenv